Разбивка на модули исправление замечаний
