+++ public/index.html
@@ -35,6 +35,6 @@
   </section>
 </footer>
 
-<script src="./bundle.js"></script>
+<script src="bundle.js"></script>
 </body>
 </html>
