+++ src/components/films-extra.js
@@ -1,0 +1,4 @@
+export const createTemplateFilmsExtra = () =>
+  `<section class="films-list--extra">
+      <div class="films-list__container"></div>
+      </section>`;
