+++ src/components/films-most-commented.js
@@ -1,0 +1,2 @@
+export const createTemplateFilmsMostCommented = () =>
+  `<h2 class="films-list__title">Most commented</h2>`;
