+++ src/components/films-top-rated.js
@@ -1,0 +1,2 @@
+export const createTemplateFilmsTopRated = () =>
+  `<h2 class="films-list__title">Top rated</h2>`;
