+++ src/components/films.js
@@ -1,19 +1,1 @@
 export const createTemplateFilmsBoard = () => `<section class="films"></section>`;
-
-export const createTemplateFilmsList = () => `<section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-
-      <div class="films-list__container"></div>
-      </section>`;
-
-export const createTemplateFilmsExtra = () =>
-  `<section class="films-list--extra">
-      <div class="films-list__container"></div>
-      </section>`;
-
-export const createTemplateFilmsTopRated = () =>
-  `<h2 class="films-list__title">Top rated</h2>`;
-
-export const createTemplateFilmsMostCommented = () =>
-  `<h2 class="films-list__title">Most commented</h2>`;
-
